Lint ruff + intégration continue GitHub Actions
CI sur windows-latest (seule plateforme supportée), déclenchée sur push et
PR vers main :
- job lint : ruff seul, sans installer Qt ;
- job tests : dépendances complètes (ce qui valide au passage que
  requirements.txt s'installe sur une machine neuve), pytest avec
  couverture, puis construction de l'interface hors écran.

Ajoute scripts/verifier_ui.py : instancie réellement MainWindow en mode
offscreen. Ce garde-fou rattrape les imports cassés et les signaux
connectés à un slot disparu, hors de portée du harnais core/.

Lint : ruff.toml (py312, ligne à 90) et mise en conformité du code —
imports triés, imports morts retirés, boucles et conditions simplifiées,
lignes longues coupées. Les règles qui contredisent les conventions du
projet sont désactivées avec leur justification : replis silencieux
voulus (S110/S112/SIM105), typographie française (RUF001-003), signaux
camelCase imposés par le contrat QWebChannel avec map.html (N802/N815
sur ui/map_panel.py), import tardif structurel de main.py (E402).

Aucun changement de comportement : les 190 tests passent et la fenêtre se
construit toujours.

+++ tests/conftest.py
@@ -47,7 +47,7 @@ def _dms_rational(value: float) -> tuple:
     minutes_float = (value - degrees) * 60
     minutes = int(minutes_float)
     seconds = (minutes_float - minutes) * 60
-    return ((degrees, 1), (minutes, 1), (int(round(seconds * 10000)), 10000))
+    return ((degrees, 1), (minutes, 1), (round(seconds * 10000), 10000))
 
 
 def write_photo(
